Fetch paired orthologs

This function provides ability to query InParanoiDB 9 and OrthoMCL 7 to get paired orthologs between two species of interest.

Usage

getpairedOrthologs(
  from,
  to,
  db = c("ipdb", "orthomcl"),
  customFields = NULL,
  transform = TRUE
)

Arguments

from

ID of query organism for which orthologs are to be fetched. To view organisms indexed in the database and their respective IDs use listipdb() or listOrthomcl().

to

ID of target organism against which will be queried for orthologs. To view organisms indexed in the database and their respective IDs use listipdb() or listOrthomcl().

db

Define database to be queried. Possible values: "orthomcl","ipdb".

customFields

Additional field to be fetched from OrthoMCL 7. "primary_key" and "target_id" are mandatory fields. Additional popular fields include: "group_name","product","source_id","num_core","num_peripheral","length","sequence","taxon_name","abbreviation","core_peripheral","ec_numbers","pfam_domains" etc.For more fields refer to OrthoMCL REST query builder.

transform

Logical. In case of InParanoiDB, transform collapses the orthologs by Group IDs so that you have unique rows. For OrthoMCL, it performs the opposite function and separates the rows so that you have only one ID in each column. When set to FALSE, the function returns the ortholog query results in the raw form provided by the database.

Value

df This function returns a dataframe of orthologs and additional requested fields.

Examples

if (FALSE) { # \dontrun{
df <- getpairedOrthologs(from=1742855, to=1747281, db="orthomcl", transform = FALSE)
df <- getpairedOrthologs(from=1742855, to=1747281, db="orthomcl", 
transform = FALSE, customFields = c("primary_key","target_id","group_name"))
df <- getpairedOrthologs(from=1742855, to=1747281, db="orthomcl", transform = TRUE)
df <- getpairedOrthologs(from=423536,to=36329, transform = FALSE)
df <- getpairedOrthologs(from=423536,to=36329, transform = TRUE)
} # }